+++ tests/test_parser.py
@@ -24,7 +24,7 @@ from __future__ import unicode_literals
 import unittest
 
 from mwparserfromhell import parser
-from mwparserfromhell.nodes import Template, Text, Wikilink
+from mwparserfromhell.nodes import Tag, Template, Text, Wikilink
 from mwparserfromhell.nodes.extras import Parameter
 
 from ._test_tree_equality import TreeEqualityTestCase, wrap, wraptext
@@ -35,10 +35,12 @@ class TestParser(TreeEqualityTestCase):
 
     def test_use_c(self):
         """make sure the correct tokenizer is used"""
+        restore = parser.use_c
         if parser.use_c:
             self.assertTrue(parser.Parser()._tokenizer.USES_C)
             parser.use_c = False
         self.assertFalse(parser.Parser()._tokenizer.USES_C)
+        parser.use_c = restore
 
     def test_parsing(self):
         """integration test for parsing overall"""
@@ -61,6 +63,27 @@ class TestParser(TreeEqualityTestCase):
         ])
         actual = parser.Parser().parse(text)
         self.assertWikicodeEqual(expected, actual)
+
+    def test_skip_style_tags(self):
+        """test Parser.parse(skip_style_tags=True)"""
+        def test():
+            with_style = parser.Parser().parse(text, skip_style_tags=False)
+            without_style = parser.Parser().parse(text, skip_style_tags=True)
+            self.assertWikicodeEqual(a, with_style)
+            self.assertWikicodeEqual(b, without_style)
+
+        text = "This is an example with ''italics''!"
+        a = wrap([Text("This is an example with "),
+                  Tag(wraptext("i"), wraptext("italics"), wiki_markup="''"),
+                  Text("!")])
+        b = wraptext("This is an example with ''italics''!")
+
+        restore = parser.use_c
+        if parser.use_c:
+            test()
+            parser.use_c = False
+        test()
+        parser.use_c = restore
 
 if __name__ == "__main__":
     unittest.main(verbosity=2)